core/src/main/java/com/redhat/osas/ml/analysis/SentimentAnalyzer.java
package com.redhat.osas.ml.analysis;

import edu.stanford.nlp.ling.CoreAnnotations;
import edu.stanford.nlp.ling.HasWord;
import edu.stanford.nlp.ling.TaggedWord;
import edu.stanford.nlp.pipeline.Annotation;
import edu.stanford.nlp.pipeline.StanfordCoreNLP;
import edu.stanford.nlp.process.DocumentPreprocessor;
import edu.stanford.nlp.rnn.RNNCoreAnnotations;
import edu.stanford.nlp.sentiment.SentimentCoreAnnotations;
import edu.stanford.nlp.tagger.maxent.MaxentTagger;
import edu.stanford.nlp.trees.Tree;
import edu.stanford.nlp.util.CoreMap;

import java.io.StringReader;
import java.util.List;
import java.util.Properties;

public class SentimentAnalyzer {
    public boolean hasSubject(String corpus, String text) {
        MaxentTagger tagger = new MaxentTagger("models/wsj-0-18-bidirectional-nodistsim.tagger");
        StringReader reader = new StringReader(corpus);
        DocumentPreprocessor preprocessor = new DocumentPreprocessor(reader);
        for (List<? extends HasWord> sentence : preprocessor) {

            // we're looking for text as a noun in the corpus.
            List<TaggedWord> words = tagger.tagSentence(sentence);
            System.out.println(words);
            for (TaggedWord w : words) {
                if (w.word().equalsIgnoreCase(text) && w.tag().startsWith("NN")) {
                    return true;
                }
            }
        }
        return false;
    }

    int findSentiment(String text) {
        Properties props = new Properties();
        props.setProperty("annotators", "tokenize, ssplit, parse, sentiment");
        StanfordCoreNLP pipeline = new StanfordCoreNLP(props);
        int mainSentiment = 0;
        if (text != null && text.length() > 0) {
            int longest = 0;
            Annotation annotation = pipeline.process(text);
            for (CoreMap sentence : annotation.get(CoreAnnotations.SentencesAnnotation.class)) {
                Tree tree = sentence.get(SentimentCoreAnnotations.AnnotatedTree.class);
                int sentiment = RNNCoreAnnotations.getPredictedClass(tree);
                String partText = sentence.toString();
                if (partText.length() > longest) {
                    mainSentiment = sentiment;
                    longest = partText.length();
                }

            }
        }
        return mainSentiment;
    }
}
